# docs/openapi.yaml
openapi: 3.1.0
info:
  title: S-SPARC AI & E-STRANGE Platform - API Reference
  version: 3.0.0
  summary: Sustainable Smart Personal Assistant for Responsible Consumption & Code Evaluation Microservice
  description: "# S-SPARC AI & E-STRANGE Platform Documentation\n\nWelcome to the official developer documentation and API reference for **S-SPARC (Sustainable Smart Personal Assistant for Responsible Consumption)** integrated with the **E-STRANGE Learning Management Platform**.\n\n---\n\n## System Architecture Overview\n\nS-SPARC is engineered around 4 core operational tiers designed for high-concurrency academic software engineering tutoring:\n\n1. **E-STRANGE Parent LMS (PHP Platform)**:\n   - Handles student course enrollment, assignment submissions, automated similarity analysis (`suspicion`), peer reviews (`code_clarity_suggestion`), gamification leaderboards, and grade tracking.\n2. **S-SPARC AI Frontend (PHP/JS Client)**:\n   - Interactive eco-aware coding workspace featuring token usage gauges, real-time AI responses, and environmental footprint visualizers.\n3. **S-SPARC FastAPI Backend (Python Core)**:\n   - Core API layer delivering user session auth, semantic search cache (0\
...
  contact:
    name: S-SPARC Engineering & Research Team
    email: support@s-sparc.ac.id
    url: http://localhost:5000/redocly
  license:
    name: Academic Research & Sustainability License 1.0
    url: https://opensource.org/licenses/MIT
servers:
- url: http://127.0.0.1:5000
...
paths:
...
  /courses:
    get:
      tags:
      - Domain/Learning
      summary: List Enrolled Courses
      description: Retrieves enrolled academic courses for the student.
      operationId: listCourses
      security:
      - SessionCookieAuth: []
      - HeaderUserIdAuth: []
      responses:
        '200':
          description: List of courses.
          content:
            application/json:
              example:
                courses:
                - course_id: CIVIL-101
                  name: Mekanika Tanah & Hidrologi
                  description: Mata kuliah teknik sipil dasar.
  /assessments:
    get:
      tags:
      - Domain/Learning
      summary: List Course Assessments
      description: Retrieves active coding assignments and submission windows.
      operationId: listAssessments
      security:
      - SessionCookieAuth: []
      - HeaderUserIdAuth: []
      parameters:
      - name: course_id
        in: query
        required: false
        schema:
          type: string
      responses:
        '200':
          description: List of active assessments.
...
components:
  securitySchemes:
    SessionCookieAuth:
      type: apiKey
      in: cookie
      name: session
      description: HTTP-only session cookie established via POST /login.
    HeaderUserIdAuth:
      type: apiKey
      in: header
      name: X-User-ID
      description: Explicit user ID header.
...